% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bornes}
\alias{bornes}
\title{auxiliary function (used in function 3 "DF_qrt") which allows to calculate the limits of a quantile}
\usage{
bornes(x, p.inf, p.sup)
}
\value{
quantile(x,probs=c(p.inf,p.sup))
}
\description{
auxiliary function (used in function 3 "DF_qrt") which allows to calculate the limits of a quantile
}
