% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DF_cl}
\alias{DF_cl}
\title{function 5 is used to modify the elements of a column of a data frame}
\usage{
DF_cl(v1, v2, data_col, data)
}
\value{
data
}
\description{
function 5 is used to modify the elements of a column of a data frame
}
