% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DF_entendue}
\alias{DF_entendue}
\title{function 1 calculates the difference between the max and the min of each column of data.frame}
\usage{
DF_entendue(data)
}
\description{
function 1 calculates the difference between the max and the min of each column of data.frame
}
