% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DF_qrt}
\alias{DF_qrt}
\title{function 3 calculates the quantile bounds for each column of data.frame}
\usage{
DF_qrt(data)
}
\description{
function 3 calculates the quantile bounds for each column of data.frame
}
