% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{DF_test_inf}
\alias{DF_test_inf}
\title{function 2 tests are the elements of column i less than the elements of column j}
\usage{
DF_test_inf(data, i, j)
}
\description{
function 2 tests are the elements of column i less than the elements of column j
}
