% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_Age}
\alias{Diab_Age}
\title{function 10 returns a data frame contains the number of women with diabetes over 30 years old and the number
women who have diabetes under 30 years old}
\usage{
Diab_Age(data)
}
\description{
function 10 returns a data frame contains the number of women with diabetes over 30 years old and the number
women who have diabetes under 30 years old
}
