% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_BP}
\alias{Diab_BP}
\title{function 9 returns a data frame containing the number of women with diabetes with abnormal diastolic blood pressure and the number
women who have diabetes with naormal diastolic blood pressure}
\usage{
Diab_BP(data)
}
\description{
function 9 returns a data frame containing the number of women with diabetes with abnormal diastolic blood pressure and the number
women who have diabetes with naormal diastolic blood pressure
}
