% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_FGD}
\alias{Diab_FGD}
\title{function 13 returns a data frame contains the number of women who have diabetes with FGD <0.5 and the number
women with diabetes with FGD> 0.5}
\usage{
Diab_FGD(data)
}
\description{
function 13 returns a data frame contains the number of women who have diabetes with FGD <0.5 and the number
women with diabetes with FGD> 0.5
}
