% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_G3}
\alias{Diab_G3}
\title{function 8 returns a data frame contains the number of women who have diabetes and were pregnant more than three times and the number
women who have diabetes but were less than three times pregnant}
\usage{
Diab_G3(data)
}
\description{
function 8 returns a data frame contains the number of women who have diabetes and were pregnant more than three times and the number
women who have diabetes but were less than three times pregnant
}
