% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_GL}
\alias{Diab_GL}
\title{function 11 returns a data frame contains the number of women who have diabetes with abnormal glucose levels and the number
of women who have diabetes with normal glucose levels}
\usage{
Diab_GL(data)
}
\description{
function 11 returns a data frame contains the number of women who have diabetes with abnormal glucose levels and the number
of women who have diabetes with normal glucose levels
}
