% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_IMC}
\alias{Diab_IMC}
\title{function 12 returns a data.frame contains the number of women who have diabetes with normal BMI and the number
women with diabetes with abnormal BMI}
\usage{
Diab_IMC(data)
}
\description{
function 12 returns a data.frame contains the number of women who have diabetes with normal BMI and the number
women with diabetes with abnormal BMI
}
