% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_MG}
\alias{Diab_MG}
\title{function 14 returns a data frame contains the number of women who have diabetes with normal skin thickness and the number
women who have diabetes with abnormal skin thickness}
\usage{
Diab_MG(data)
}
\description{
function 14 returns a data frame contains the number of women who have diabetes with normal skin thickness and the number
women who have diabetes with abnormal skin thickness
}
