% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Diab_test}
\alias{Diab_test}
\title{function 7 calculates the number of women who have the positive "diabetes" test and the number of women who have the negative "diabetes" test}
\usage{
Diab_test(data)
}
\description{
function 7 calculates the number of women who have the positive "diabetes" test and the number of women who have the negative "diabetes" test
}
