% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes-package.R
\docType{package}
\name{diabetes-package}
\alias{diabetes-package}
\alias{diabetes}
\title{Analyze Diabetes Patients}
\description{
The description. It needs to contain at least two sentences.
}
\details{
\tabular{ll}{
Package: \tab diabetes\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-06-08\cr
License: \tab LGPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Author Name <author@name.com>
}
\keyword{package}
