% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{get_patient_duration}
\alias{get_patient_duration}
\title{Distribution of duration of study participation}
\usage{
get_patient_duration(df)
}
\arguments{
\item{df}{Diabetes data frame produced by \link{read_diabetes}}
}
\description{
Distribution of duration of study participation
}
\examples{
\dontrun{
hist(get_patient_duration(df))
}
}
