% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{info}
\alias{info}
\title{function 6 takes a vector as argument and returns the associated histogram where four vertical lines are drawn (min, median, mean, max)+the density function curve}
\usage{
info(data_col)
}
\description{
function 6 takes a vector as argument and returns the associated histogram where four vertical lines are drawn (min, median, mean, max)+the density function curve
}
