% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{normalize_times}
\alias{normalize_times}
\title{Normalize times to a standard set}
\usage{
normalize_times(df)
}
\arguments{
\item{df}{Diabetes data frame produced by \link{read_diabetes}}
}
\description{
Normalize times to a standard set
}
