% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_diabetes}
\alias{read_diabetes}
\title{Read the diabetes dataset}
\usage{
read_diabetes(bad = c(2, 27, 29, 40), base = "private/diabetes")
}
\arguments{
\item{bad}{Files with bad data}

\item{base}{The base path of the diabetes files}
}
\value{
A data.frame of the merged files
}
\description{
Reads in all the individual patient files and consolidates into a single
data frame.
}
